requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==18.9b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Django==2.1.7
django-cors-headers==2.5.0
django-templated-mail==1.1.1
djangorestframework==3.9.2
djoser==1.4.0
html5lib==1.0.1
idna==2.8
numpy==1.16.2
pandas==0.23.4
Pillow==5.4.1
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2018.9
reportlab==3.5.13
requests==2.21.0
six==1.12.0
toml==0.10.0
urllib3==1.24.1
webencodings==0.5.1
xhtml2pdf==0.2.3
